Changing the default connection ports for nodes + RPC ports (MAINNET and TESTNET)
The changes made are intended to reduce the number of existing conflicts with other wallets and/or programs that use the current ports by default.
The number of each door refers to the number of this letter in the alphabet, according to the numerological correspondence calculation.
M = 4
I = 9
C = 3
M(ain) = 4
N(ode) = 5